# docs/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/rest/v1/sheets/{id}/shots": {
      "get": {
        "description": "This will return every shot for the sheet with the given id, as a JSON\narray with the same shape as GET /rest/v1/shots. Returns an empty array\nfor an existing sheet without shots.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Get shots by sheet id",
        "operationId": "getShotsBySheetId",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the sheet whose shots to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    },
    "/rest/v1/shots": {
      "get": {
        "description": "This will show all shots by default.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "shots"
        ],
        "summary": "Get all shots",
        "operationId": "getAllShots",
        "responses": {
          "200": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "post": {
        "description": "This will create a new shot.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "shots"
        ],
        "summary": "Create shots",
        "operationId": "createShot",
        "parameters": [
          {
            "description": "The request body for creating a shot",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateShotRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    },
    "/rest/v1/shots/{id}": {
      "get": {
        "description": "This will get the shot with the given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "shots"
        ],
        "summary": "Get shots",
        "operationId": "getShot",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the shot to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "put": {
        "description": "This will update a shot by its given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "shots"
        ],
        "summary": "Update shots",
        "operationId": "updateShotById",
        "parameters": [
          {
            "description": "The request body for updating a shot",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateShotByIdRequest"
            }
          },
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the shot to update",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "delete": {
        "description": "This will delete a shot by its given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "shots"
        ],
        "summary": "Delete shots",
        "operationId": "deleteShot",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the shot to delete",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ItemDeletedResponse represents the response when an item is deleted",
            "schema": {
              "$ref": "#/definitions/ItemDeletedResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "ComparisonWithPreviousResult": {
      "description": "0 = worst, 1 = same, 2 = better, 3 = unknown.",
      "type": "integer",
      "format": "uint8",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/models/sql"
    },
...
    "CreateShotRequest": {
      "description": "CreateShotRequest represents the request body for creating a shot",
      "type": "object",
      "properties": {
        "additional_notes": {
          "type": "string",
          "x-go-name": "AdditionalNotes"
        },
        "beans_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "BeansId"
        },
        "comparison_with_previous_result": {
          "$ref": "#/definitions/ComparisonWithPreviousResult"
        },
        "grind_setting": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "GrindSetting"
        },
        "is_too_bitter": {
          "type": "boolean",
          "x-go-name": "IsTooBitter"
        },
        "is_too_sour": {
          "type": "boolean",
          "x-go-name": "IsTooSour"
        },
        "quantity_in": {
          "type": "number",
          "format": "double",
          "x-go-name": "QuantityIn"
        },
        "quantity_out": {
          "type": "number",
          "format": "double",
          "x-go-name": "QuantityOut"
        },
        "rating": {
          "type": "number",
          "format": "double",
          "x-go-name": "Rating"
        },
        "sheet_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SheetId"
        },
        "shot_time": {
          "$ref": "#/definitions/Duration"
        },
        "water_temperature": {
          "type": "number",
          "format": "double",
          "x-go-name": "WaterTemperature"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    },
    "Duration": {
      "description": "A Duration represents the elapsed time between two instants\nas an int64 nanosecond count. The representation limits the\nlargest representable duration to approximately 290 years.",
      "type": "integer",
      "format": "int64",
      "x-go-package": "time"
    },
    "ItemDeletedResponse": {
      "description": "ItemDeletedResponse represents the response when an item is deleted",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id"
        },
        "msg": {
          "type": "string",
          "x-go-name": "Msg"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    },
...
    "UpdateShotByIdRequest": {
      "description": "UpdateShotByIdRequest represents the request body for updating a shot\nwith the given id",
      "type": "object",
      "properties": {
        "additional_notes": {
          "type": "string",
          "x-go-name": "AdditionalNotes"
        },
        "beans_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "BeansId"
        },
        "comparison_with_previous_result": {
          "$ref": "#/definitions/ComparisonWithPreviousResult"
        },
        "grind_setting": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "GrindSetting"
        },
        "is_too_bitter": {
          "type": "boolean",
          "x-go-name": "IsTooBitter"
        },
        "is_too_sour": {
          "type": "boolean",
          "x-go-name": "IsTooSour"
        },
        "quantity_in": {
          "type": "number",
          "format": "double",
          "x-go-name": "QuantityIn"
        },
        "quantity_out": {
          "type": "number",
          "format": "double",
          "x-go-name": "QuantityOut"
        },
        "rating": {
          "type": "number",
          "format": "double",
          "x-go-name": "Rating"
        },
        "sheet_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SheetId"
        },
        "shot_time": {
          "$ref": "#/definitions/Duration"
        },
        "water_temperature": {
          "type": "number",
          "format": "double",
          "x-go-name": "WaterTemperature"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    }
  },
  "responses": {
...
    "ErrorResponse": {
      "description": "ErrorResponse represents the json response\nfor http errors.\nIt contains a message describing the error",
      "headers": {
        "msg": {
          "type": "string"
        }
      }
    },
...
    "ShotResponse": {
      "description": "ShotResponse represents an espresso shot for this application\n\nAn espresso shot is made from coffee beans, ground at a specific setting,\nwith a specific quantity of coffee in and out.\nIt also has a specific shot time and water temperature.\n\nThe result of a shot can be rated and compared to the previous shot.\nIt can also be too bitter or too sour.",
      "headers": {
        "additional_notes": {
          "type": "string"
        },
        "beans": {},
        "comparison_with_previous_result": {
          "type": "integer",
          "format": "uint8"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "grind_setting": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "is_too_bitter": {
          "type": "boolean"
        },
        "is_too_sour": {
          "type": "boolean"
        },
        "quantity_in": {
          "type": "number",
          "format": "double"
        },
        "quantity_out": {
          "type": "number",
          "format": "double"
        },
        "rating": {
          "type": "number",
          "format": "double"
        },
        "sheet": {},
        "shot_time": {
          "type": "integer",
          "format": "int64"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "water_temperature": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}